requirements.txt
pypubsub
pyyaml